VitalControllerのリファクタリング
