+++ app/UseCases/Vital/VitalDeleteUseCase.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\UseCases\Vital;
+
+use App\Models\Repositories\VitalRepository;
+
+final class VitalDeleteUseCase
+{
+    /**
+     * @var App\Models\Repositories\VitalRepository
+     */
+    private VitalRepository $repository;
+
+    public function __construct(VitalRepository $repository)
+    {
+        $this->repository = $repository;
+    }
+
+    public function execute($request)
+    {
+        return $this->repository->delete($request);
+    }
+}
